split collection-add-to comp. for flexibility > collection-add-to-list
